com.unity.addressables.android@1.0.6 ## [1.0.6] - 2025-02-12
* Loading Addressable assets synchronously can now generate an exception. For more information, refer to the documentation.
* Updated `com.unity.addressables` dependency up to v2.3.16.

+++ Tests/Runtime/PlayAssetDeliveryTestsBase.cs
@@ -387,9 +387,12 @@ internal abstract class PlayAssetDeliveryTestsBase
         Assert.NotNull(emptyFolderEntry);
     }
 
-    protected void CleanupAddressables()
+    protected void CleanupAddressables(bool deleteServerData)
     {
-        DeleteDirectoryFromAssets("ServerData");
+        if (deleteServerData)
+        {
+            DeleteDirectoryFromAssets("ServerData");
+        }
         DeleteDirectoryFromAssets(Path.Combine("Assets", kTestFolder));
         DeleteDirectoryFromAssets(CustomAssetPackUtility.RootDirectory);
         DeleteDirectoryFromAssets(AddressableAssetSettingsDefaultObject.kDefaultConfigFolder);
@@ -417,7 +420,7 @@ internal abstract class PlayAssetDeliveryTestsBase
     [TearDown]
     public void CleanupEditorTests()
     {
-        CleanupAddressables();
+        CleanupAddressables(true);
     }
 #endif
 }
